split MockCompilerFactory to a separate plugin
