+++ org.eclipse.jdt.core.tests.builder.mockcompiler/.settings/org.eclipse.core.runtime.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+line.separator=\n
