+++ org.eclipse.jdt.core.tests.builder.mockcompiler/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %pluginName
+Bundle-SymbolicName: org.eclipse.jdt.core.tests.builder.mockcompiler; singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: %providerName
+Bundle-Localization: plugin
+Fragment-Host: org.eclipse.jdt.core
+Export-Package: org.eclipse.jdt.core.tests.builder.mockcompiler
+Require-Bundle: org.eclipse.jdt.core;bundle-version="[3.36.0,4.0.0)"
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Eclipse-BundleShape: dir
+Automatic-Module-Name: org.eclipse.jdt.core.tests.builder.mockcompiler
