+++ org.eclipse.jdt.core.tests.builder.mockcompiler/build.properties
@@ -1,7 +1,6 @@
 ###############################################################################
-# Copyright (c) 2000, 2013 IBM Corporation and others.
-#
-# This program and the accompanying materials
+# Copyright (c) 2024 Microsoft Corporation and others.
+# All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
 # https://www.eclipse.org/legal/epl-2.0/
@@ -9,16 +8,12 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 # Contributors:
-#     IBM Corporation - initial API and implementation
-#     Red Hat Inc. (mistria) - Avoid nested jars
+#     Microsoft Corporation - initial API and implementation
 ###############################################################################
-bin.includes = plugin.xml,\
-               test.xml,\
-               about.html,\
+bin.includes = about.html,\
                .,\
                META-INF/,\
-               plugin.properties,\
-               workspace/
+               plugin.properties
 source.. = src/
 output.. = bin/
 src.includes=about.html
