+++ org.eclipse.jdt.core.tests.builder.mockcompiler/plugin.properties
@@ -1,0 +1,14 @@
+###############################################################################
+# Copyright (c) 2024 Microsoft Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# https://www.eclipse.org/legal/epl-2.0/
+#
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#     Microsoft Corporation - initial API and implementation
+###############################################################################
+providerName=Eclipse.org
+pluginName=Java Custom Compiler Tests
