+++ org.eclipse.jdt.core.tests.builder/plugin.xml
